♻️ [refactor]: #452 Phase 16 (G): lexer_next_token.rs を分離
next_token の 30 テスト（Phase G）を
`src/lexer/tests/lexer_next_token.rs` に分離。テスト本体で
`Token::Primitive(Primitive::Name(PdfName::new(...)))` の形で
`PdfName::new` を直接参照するため
`use crate::object::name::PdfName;` を追加。

tests.rs の最後のテスト本体を移動完了。不要となった
`use super::*;` を削除し、最終状態は 16 個の `mod lexer_*;`
宣言のみとなった。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_next_token.rs: 246 行 (< 800)
- tests.rs: 16 行 (mod 宣言のみ)

Refs #452
